- fixed subwindows sizes: now these are resized to the workspace size - tableview's cells text positioning fixed, synced with editors,   cells sizes fixed, numeric cells' editors are now right-aligned
svn path=/trunk/koffice/; revision=231450

+++ kexi/plugins/queries/kexiquerydesignersqlhistory.cpp
@@ -214,7 +214,7 @@ HistoryEntry::highlight()
 	statement.replace(QRegExp("\t"), "&nbsp;&nbsp;&nbsp;");
 #endif
 	// getting quoting...
-	for(int i=0; i < statement.length(); i++)
+	for(int i=0; i < (int)statement.length(); i++)
 	{
 		QString beginTag;
 		QString endTag;
